Release 4.2.0
Release notes are available at:
https://docs.dremio.com/release-notes/42-release-notes.html

+++ common/src/main/java/com/dremio/common/memory/AllocatorUtil.java
@@ -30,7 +30,7 @@ public class AllocatorUtil {
    * @param allocator
    * @param headRoom
    */
-  public static void ensureHeadroom(BufferAllocator allocator, int headRoom) throws OutOfMemoryException {
+  public static void ensureHeadroom(BufferAllocator allocator, long headRoom) throws OutOfMemoryException {
     if (allocator.getHeadroom() >= headRoom) {
       // nothing to do in this case.
       return;
