+++ sabot/kernel/src/main/java/com/dremio/exec/expr/fn/impl/conv/BigIntVLongConvertTo.java
@@ -54,6 +54,6 @@ public class BigIntVLongConvertTo implements SimpleFunction {
     com.dremio.exec.util.ByteBufUtil.HadoopWritables.writeVLong(errorContext, buffer, 0, 9, in.value);
     out.buffer = buffer;
     out.start = 0;
-    out.end = buffer.readableBytes();
+    out.end = (int) buffer.readableBytes();
   }
 }
