+++ sabot/kernel/src/main/java/com/dremio/sabot/op/sender/roundrobin/RoundRobinOperator.java
@@ -160,7 +160,7 @@ public class RoundRobinOperator extends BaseSender {
         @Nullable
         @Override
         public ArrowBuf apply(@Nullable ArrowBuf buf) {
-          int writerIndex = buf.writerIndex();
+          long writerIndex = buf.writerIndex();
           ArrowBuf newBuf = buf.getReferenceManager().transferOwnership(buf, allocator).getTransferredBuffer();
           newBuf.writerIndex(writerIndex);
           buf.release();
